Format code

+++ include/battle_bg.h
@@ -6,17 +6,17 @@
 
 struct BattleBackground
 {
-	const u8 *name;
+    const u8 *name;
     const void *tileset;
     const void *tilemap;
     const void *entryTileset;
     const void *entryTilemap;
     const void *palette;
-	TaskFunc introSlideFunc;
-	u8 secretPowerEffect;
-	u8 camouflageType;
-	u16 camouflageBlendColor;
-	u16 naturePowerMove;
+    TaskFunc introSlideFunc;
+    u8 secretPowerEffect;
+    u8 camouflageType;
+    u16 camouflageBlendColor;
+    u16 naturePowerMove;
 };
 
 extern const struct BgTemplate gBattleBgTemplates[];
@@ -25,7 +25,7 @@ extern const struct BattleBackground gBattleTerrainTable[BATTLE_TERRAINS_COUNT];
 void LoadBattleTerrainGfx(u32 terrain);
 void InitBattleBgsVideo(void);
 void DrawBattleEntryBackground(void);
-void InitLinkBattleVsScreen(u8 taskId);
+void InitLinkBattleVsScreen(u32 taskId);
 void LoadBattleMenuWindowGfx(void);
 void LoadBattleTextboxAndBackground(void);
 void BattleInitBgsAndWindows(void);
