+++ include/trainer_card.h
@@ -49,8 +49,8 @@ struct TrainerCard
 extern struct TrainerCard gTrainerCards[4];
 
 void ShowPlayerTrainerCard(void (*callback)(void));
-void ShowTrainerCardInLink(u8 whoseCard, MainCallback callback);
-u8 GetTrainerCardStars(u8 cardId);
+void ShowTrainerCardInLink(u32 whoseCard, MainCallback callback);
+u32 GetTrainerCardStars(u32 cardId);
 void TrainerCard_GenerateCardForLinkPlayer(struct TrainerCard *trainerCard);
 
 #endif //GUARD_TRAINER_CARD_H
